% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nmb_analysis}
\alias{is.nmb_analysis}
\title{Is a Nimble Analysis}
\usage{
is.nmb_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'nmb_analysis'
}
