% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nmb_code}
\alias{is.nmb_code}
\title{Is nimble Code}
\usage{
is.nmb_code(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'nmb_code'
}
