% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nmb_model}
\alias{is.nmb_model}
\title{Is a nimble Model}
\usage{
is.nmb_model(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'nmb_model'
}
