% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbr-package.R
\docType{package}
\name{nmbr-package}
\alias{nmbr}
\alias{nmbr-package}
\title{nmbr: Analyses using nimble}
\description{
Facilitates analyses using nimble package. Part of the embr family of packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/nmbr}
  \item Report bugs at \url{https://github.com/poissonconsulting/nmbr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@pissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [funder, copyright holder]
}

}
\keyword{internal}
